moved models from game_service to matchMaking
